AnalyzersModule.  Registers AnalyzerRunner
